kunnskap/requirements.txt
Flask==2.3.3
gunicorn==21.2.0
mysql-connector-python==8.0.33
python-dotenv==1.0.1